+++ verif/ovip_ace/src/ovip_ace_agent.sv
@@ -24,6 +24,8 @@ class ovip_ace_agent extends uvm_agent;
 	ovip_axi_base_driver#(virtual ovip_ace_agent_if) drv;
 	ovip_axi_base_sequencer                          sqr;
 
+	ovip_axi_trans_logger                            trans_logger; // when cfg.enable_trans_log
+
 	`uvm_component_utils(ovip_ace_agent)
 
 	function new(string name = "ovip_ace_agent", uvm_component parent);
@@ -54,6 +56,16 @@ function void ovip_ace_agent::build_phase(uvm_phase phase);
 
 	mon = ovip_ace_monitor::type_id::create("mon", this);
 	mon.cfg = cfg;
+
+	if(cfg.enable_trans_log)
+	begin
+		string base = (cfg.agent_tag != "") ? cfg.agent_tag : get_name();
+		trans_logger = ovip_axi_trans_logger::type_id::create("trans_logger", this);
+		trans_logger.label              = base;
+		trans_logger.file_name          = (cfg.trans_log_file != "") ? cfg.trans_log_file : $sformatf("%s_trans.log", base);
+		trans_logger.combined_file_name = cfg.trans_log_combined_file;
+		trans_logger.format             = cfg.trans_log_format;
+	end
 
 	if(get_is_active())
 	begin
@@ -90,6 +102,9 @@ function void ovip_ace_agent::connect_phase(uvm_phase phase);
 	vif.is_active = (get_is_active());
 
 	mon.vif = vif;
+
+	if(trans_logger != null)
+		mon.analysis_port.connect(trans_logger.analysis_export);
 
 	if(get_is_active())
 	begin